Improve DX for PaginationExtension reusability
